Integrated plg.py and plg_Tpred.py
